requirements.txt
Django==3.2.12
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-environ==0.8.1
drf-yasg==1.20.0
cryptography==36.0.2
gunicorn==20.1.0
